refactor: Move DocFormatter to main from test sources
