+++ src/test/java/org/opentripplanner/generate/doc/framework/OTPFeatureTable.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.framework.doc.DocFormatter;
 import org.opentripplanner.util.OTPFeature;
 import org.opentripplanner.util.lang.TableFormatter;
